add card.ts
